fix(BAE-22): Fixed DELETE endpoints
