+++ Backend/requirements.txt
@@ -1,24 +1,8 @@
 anyio==3.6.2
-asgiref==3.5.0
-attrs==22.1.0
+certifi==2022.9.24
 boto3==1.26.15
 botocore==1.29.15
-certifi==2021.10.8
-cffi==1.15.1
-charset-normalizer==2.0.12
 click==8.1.3
-colorama==0.4.6
-crispy-bootstrap5==0.6
-cryptography==38.0.1
-distlib==0.3.4
-dj-database-url==1.0.0
-Django==4.0.3
-django-bootstrap-icons==0.7.8
-django-cors-headers==3.12.0
-django-crispy-forms==1.14.0
-django-heroku==0.3.1
-django-passwords==0.3.12
-djangorestframework==3.13.1
 dnspython==2.2.1
 email-validator==1.3.0
 et-xmlfile==1.1.0
@@ -37,53 +21,27 @@ Jinja2==3.1.2
 jmespath==1.0.1
 MarkupSafe==2.1.1
 mysql-connector-python==8.0.31
-numpy==1.23.4
-openai==0.25.0
-openpyxl==3.0.10
-orjson==3.8.1
-packaging==21.3
-pandas==1.5.1
-pandas-stubs==1.5.1.221024
-pbr==5.9.0
-Pillow==9.2.0
-platformdirs==2.5.2
-pluggy==1.0.0
+numpy==1.23.5
+pandas==1.5.2
 protobuf==3.20.1
-psycopg2==2.9.3
-pycparser==2.21
 pydantic==1.10.2
 PyMySQL==1.0.2
 PyMysqlDB==0.0.2
-pyparsing==3.0.9
-pytest==7.2.0
 python-dateutil==2.8.2
-python-decouple==3.6
 python-dotenv==0.21.0
 python-multipart==0.0.5
-pytz==2022.1
+pytz==2022.6
 PyYAML==6.0
-requests==2.27.1
 rfc3986==1.5.0
 s3transfer==0.6.0
 six==1.16.0
 sniffio==1.3.0
+stevedore==4.0.0
 sql==2022.4.0
 SQLAlchemy==1.4.44
-sqlparse==0.4.2
 starlette==0.21.0
-stevedore==4.0.0
-tomli==2.0.1
-tqdm==4.64.1
-types-pytz==2022.6.0.1
-typing_extensions==4.4.0
-tzdata==2022.1
-ujson==5.5.0
-urllib3==1.26.9
+typing-extensions==4.4.0
 uvicorn==0.19.0
-virtualenv==20.14.1
-virtualenv-clone==0.5.7
-virtualenvwrapper==4.8.4
-virtualenvwrapper-win==1.2.7
+uvloop==0.17.0
 watchfiles==0.18.1
 websockets==10.4
-whitenoise==6.2.0
