+++ Backend/tests/test_delete_statement.py
@@ -1,11 +1,20 @@
 from fastapi.testclient import TestClient
-from routes.delete_documents import router
+from ..routes.delete_documents import router
 
 client = TestClient(router)
 
 def test_delete_statement_success():
     """ Tests a successful delete """
     url = 'http://127.0.0.1:8000/delete_statement/{filename}'
-    file = {'file': open('dummy_account_statement.csv', 'rb')}
-    response = client.delete(url=url, files=file)
+    file = open(__file__, 'r+')
+    response = client.delete(url=url)
     assert response.status_code == 204
+
+def test_delete_statement_record():
+    """ Tests a successful delete """
+    url = 'http://127.0.0.1:8000/delete_record/{filename}'
+    file = open(__file__, 'r+')
+    response = client.delete(url=url)
+    assert response.status_code == 204
+
+
